get_wndprb

Return dataframe of wind speed probability data.

get_wndprb(link)

Arguments

link
URL to storm's archive page.

Source

http://www.nhc.noaa.gov/about/pdf/About_Windspeed_Probabilities.pdf

Details

Wind Speed Probability product replaced Strike Probabilities product after the 2005 hurricane season. These products may not be issued for every advisory/cyclone.

Status

Classification of storm, e.g., Tropical Storm, Hurricane, etc.

Name

Name of storm

Adv

Advisory Number

Date

Date of advisory issuance

Wind

Minimum wind speed for which probabilities reference

Wind12

Probability of sustained `Wind` within 12 hours

Wind24

Probability of sustained `Wind` within 24 hours

Wind24Cum

Cumulative probability through 24 hours

Wind36

Probability of sustained `Wind` within 36 hours

Wind36Cum

Cumulative probability through 36 hours

Wind48

Probability of sustained `Wind` within 48 hours

Wind48Cum

Cumulative probability through 48 hours

Wind72

Probability of sustained `Wind` within 72 hours

Wind72Cum

Cumulative probability through 72 hours

Wind96

Probability of sustained `Wind` within 96 hours

Wind96Cum

Cumulative probability through 96 hours

Wind120

Probability of sustained `Wind` within 120 hours

Wind120Cum

Cumulative probability through 120 hours