tidy_wr

Tidy current wind radius of a fstadv dataframe object.

tidy_wr(df)

Arguments

df
fstadv dataframe object

Details

Returns tidy dataframe of current wind radius values for a cyclone. Returns only complete.cases (based on quadrants).

Key

Unique identifier of cyclone

Adv

Advisory number

Date

Date and time of advisory

Windfield

Minimum wind speed expected

NE

Radius of `Windfield` in the northeast quadrant

SE

Radius of `Windfield` in the southeast quadrant

SW

Radius of `Windfield` in the southwest quadrant

NW

Radius of `Windfield` in the northwest quadrant

Examples

not_run({
  get_fstadv("http://www.nhc.noaa.gov/archive/1998/1998ALEXadv.html") %>%
      tidy_wr()
})