get_fstadv

Return dataframe of forecast/advisory data.

get_fstadv(links)

Arguments

links
URL to storms' archive page.

Details

Returns a wide dataframe of most the data available in a cyclones forecast/advisory product (watches and warnings are not included at this time).

Overall structure of the dataframe is listed below. Note the following clarifications:

The value of `n` in `Hr{n}` variables is the forecast period. Up to 2002, forecast periods are 12, 24, 36, 48 and 72 hours. After 2002, forecast periods were extended to 96 and 120 hours. Not all forecast periods will be available for every cyclone advisory (e.g., if it is dissipating or expected to dissipate.)

Wind radius data is not included 96 and 120 hour forecast periods.

Forecast dates are not truly 12, 24, ..., 120 hours from the date/time of the advisory. The NHC issues two positions in these products; one for current and one for three hours prior. It is the latter position the forecast date/times are based.

Status

Classification of cyclone

Name

Name of cyclone

Adv

Advisory number

Date

Date and time of advisory

Key

Unique identifier of cyclone

Lat

Latitude of cyclone center

Lon

Longitude of cyclone center

Wind

Maximum sustained one-minute winds in knots

Gust

Maximum sustained one-minute gusts in knots

Pressure

Minimum central pressure in millibars

PosAcc

Position accuracy of cyclone in nautical miles

FwdDir

Compass angle of forward motion

FwdSpeed

Forward speed in miles per hour

Eye

Size of eye in nautical miles

NE64

Radius of >=64kt winds in northeast quadrant

SE64

Radius of >=64kt winds in southeast quadrant

SW64

Radius of >=64kt winds in southwest quadrant

NW64

Radius of >=64kt winds in northwest quadrant

NE50

Radius of >=50kt winds in northeast quadrant

SE50

Radius of >=50kt winds in southeast quadrant

SW50

Radius of >=50kt winds in southwest quadrant

NW50

Radius of >=50kt winds in northwest quadrant

NE34

Radius of >=34kt winds in northwest quadrant

SE34

Radius of >=34kt winds in southeast quadrant

SW34

Radius of >=34kt winds in southwest quadrant

NW34

Radius of >=34kt winds in northwest quadrant

Hr{n}FcstDate

Forecast valid date

Hr{n}Lat

Forecast latitude in `n` hours

Hr{n}Lon

Forecast longitude in `n` hours

Hr{n}Wind

Forecast maximum wind in `n` hours

Hr{n}Gust

Forecast maximum gust in `n` hours

Hr{n}NE64

Forecast wind radius in `n` hours

Hr{n}SE64

Forecast wind radius in `n` hours

Hr{n}SW64

Forecast wind radius in `n` hours

Hr{n}NW64

Forecast wind radius in `n` hours

Hr{n}NE50

Forecast wind radius in `n` hours

Hr{n}SE50

Forecast wind radius in `n` hours

Hr{n}SW50

Forecast wind radius in `n` hours

Hr{n}NW50

Forecast wind radius in `n` hours

Hr{n}NE34

Forecast wind radius in `n` hours

Hr{n}SE34

Forecast wind radius in `n` hours

Hr{n}SW34

Forecast wind radius in `n` hours

Hr{n}NW34

Forecast wind radius in `n` hours

SeasNE

Radius of 12ft seas in northeast quadrant

SeasSE

Radius of 12ft seas in southeast quadrant

SeasSW

Radius of 12ft seas in southwest quadrant

SeasNW

Radius of 12ft seas in northwest quadrant

See also

Examples

# NOT RUN {
# Return dataframe of forecast/advisories for Tropical Storm Alex (AL011998)
get_fstadv("http://www.nhc.noaa.gov/archive/1998/1998ALEXadv.html")
# }