load_storm_data

Load storm and year data from data repository.

load_storm_data(dataset = c("adv", "discus", "fcst", "fcst_wr", "fstadv",
  "posest", "prblty", "public", "storms", "update", "wndprb", "wr"))

Arguments

dataset
A dataset to return

Details

This function is designed to give quicker access to post-scraped storm data and may be modified in future releases.

adv

This data file contains base information for every storm advisory issued by the National Hurricane Center for the Atlantic and northeast Pacific oceans.

discus

Storm discussion text

fcst

Forecast positions of tropical cyclones

fcst_wr

Forecast wind radii data for each forecast observation in `fcst`

fstadv

Complete Forecast/Advisory data

posest

Position estimates

prblty

Strike probabilities for a given location. This product was deprecated after the 2005 hurricane season.

public

Public advisory text

storms

Summary for all storms in this dataset.

update

Update text

wndprb

Wind Speed Probabilities. Probability of a location experiencing minimum wind-speed values within a given forecast period. This product replaces Strike Probabilities after the 2005 hurricane season. May not exist for all cyclones.

wr

Current wind radius, if available, for a cyclone.

Datasets 'discus', 'posest', 'public' and 'update' are not included as of this writing but will be added as soon as possible.

'adv', 'fcst', 'fcst_wr', and 'wr' are tidied data of get_fstadv using functions tidy_fcst, tidy_fcst_wr, tidy_fstadv and tidy_wr.

See https://timtrice.github.io/rrricanes/articles/articles/data_world.html for a third alternative to access datasets with the ability to filter data without loading entire datasets into your environment.

See also

https://timtrice.github.io/rrricanes/articles/articles/data_world.html